Rename Send_Item_Update to Send_Item_Updates and update docblock on Request
