+++ includes/API/Catalog/Send_Item_Updates/Response.php
@@ -8,14 +8,14 @@
  * @package FacebookCommerce
  */
 
-namespace SkyVerge\WooCommerce\Facebook\API\Catalog\Send_Item_Update;
+namespace SkyVerge\WooCommerce\Facebook\API\Catalog\Send_Item_Updates;
 
 defined( 'ABSPATH' ) or exit;
 
 use SkyVerge\WooCommerce\PluginFramework\v5_5_4 as Framework;
 
 /**
- * Base API response object
+ * Send_Item_Updates API response object
  *
  * @since 2.0.0-dev.1
  */
